refactor: Generalize `ensureXName` logic
